CassandraRepositoryGenerator non duplication for mappers R2dbcStatementSetterGenerator batch parameter fixed DbUtils#resultMapperName mapper name fixed Tests fixed

+++ database/database-annotation-processor/src/test/java/ru/tinkoff/kora/database/common/annotation/processor/jdbc/repository/AllowedResultsRepository.java
@@ -22,7 +22,7 @@ public interface AllowedResultsRepository extends JdbcRepository {
     UpdateCount returnUpdateCount();
 
     @Query("SELECT test")
-    int returnPrimitive();
+    short returnPrimitive();
 
     @Query("INSERT INTO test(test) VALUES (:someint)")
     int[] returnBatchUpdate(@Batch List<Integer> someint);
@@ -31,11 +31,11 @@ public interface AllowedResultsRepository extends JdbcRepository {
     void returnBatchVoid(@Batch List<Integer> someint);
 
     @Query("SELECT test")
-    Integer returnObject();
+    Short returnObject();
 
     @Query("SELECT test")
     @Nullable
-    Integer returnNullableObject();
+    Short returnNullableObject();
 
     @Query("SELECT test")
     @Mapping(JdbcEntity.TestEntityJdbcRowMapper.class)
